🔥 chore(examples/with-redux): overhaul redux module
