+++ examples/with-redux/src/components/Counter/index.ts
@@ -1,2 +1,1 @@
 export * from './Counter'
-export * from './counterAPI'
