+++ examples/with-redux/src/lib/redux/slices/counterSlice/fetchIdentityCount.ts
@@ -1,5 +1,7 @@
-export const fetchCount = async (amount = 1): Promise<{ data: number }> => {
-  const response = await fetch('http://localhost:3000/api/counter', {
+export const fetchIdentityCount = async (
+  amount = 1
+): Promise<{ data: number }> => {
+  const response = await fetch('http://localhost:3000/api/identity-count', {
     method: 'POST',
     headers: { 'Content-Type': 'application/json' },
     body: JSON.stringify({ amount }),
