+++ examples/with-redux/src/lib/redux/slices/counterSlice/selectors.ts
@@ -1,0 +1,7 @@
+/* Instruments */
+import type { ReduxState } from '@/lib/redux'
+
+// The function below is called a selector and allows us to select a value from
+// the state. Selectors can also be defined inline where they're used instead of
+// in the slice file. For example: `useSelector((state: RootState) => state.counter.value)`
+export const selectCount = (state: ReduxState) => state.counter.value
